Move component files to ui/
