--- packages/components/src/control-label/index.js
@@ -1,2 +1,0 @@
-export { default } from './control-label';
-export * from './use-control-label';
