+++ packages/components/src/ui/control-label/index.js
@@ -1,0 +1,2 @@
+export { default as ControlLabel } from './control-label';
+export * from './use-control-label';
