+++ packages/components/src/ui/form-group/form-group-label.js
@@ -6,8 +6,8 @@ import { memo } from '@wordpress/element';
 /**
  * Internal dependencies
  */
-import ControlLabel from '../control-label';
-import VisuallyHidden from '../visually-hidden';
+import { ControlLabel } from '../control-label';
+import VisuallyHidden from '../../visually-hidden';
 
 /**
  * @typedef OwnProps
