+++ packages/components/src/ui/form-group/index.js
@@ -1,4 +1,4 @@
-export { default } from './form-group';
+export { default as FormGroup } from './form-group';
 
 export * from './use-form-group';
 export * from './form-group-context';
