+++ packages/components/src/ui/index.js
@@ -1,4 +1,6 @@
+export * from './control-label';
 export * from './flex';
+export * from './form-group';
 export * from './grid';
 export * from './h-stack';
 export * from './text';
